Moved activity tests to a more suitable package
