+++ app/src/androidTest/kotlin/com/alancamargo/desafioinfoglobo/activities/ArticleDetailsActivityTest.kt
@@ -1,8 +1,7 @@
-package com.alancamargo.desafioinfoglobo
+package com.alancamargo.desafioinfoglobo.activities
 
 import android.content.Intent
 import androidx.test.platform.app.InstrumentationRegistry
-import com.alancamargo.desafioinfoglobo.activities.ArticleDetailsActivity
 import com.alancamargo.desafioinfoglobo.base.BaseActivityTest
 import com.alancamargo.desafioinfoglobo.model.Article
 import com.alancamargo.desafioinfoglobo.model.Image
